+++ chimeraccodecommand.cpp
@@ -299,7 +299,7 @@ int ChimeraCcodeCommand::execute(){
 		
 				int pid, numSeqsPerProcessor; 
 				int tag = 2001;
-				vector<unsigned long int> MPIPos;
+				vector<unsigned long long> MPIPos;
 								
 				MPI_Status status; 
 				MPI_Comm_rank(MPI_COMM_WORLD, &pid); //find out who we are
@@ -390,7 +390,7 @@ int ChimeraCcodeCommand::execute(){
 
 			outHeader.close();
 			
-			vector<unsigned long int> positions = m->divideFile(fastaFileNames[s], processors);
+			vector<unsigned long long> positions = m->divideFile(fastaFileNames[s], processors);
 				
 			for (int i = 0; i < (positions.size()-1); i++) {
 				lines.push_back(new linePair(positions[i], positions[(i+1)]));
@@ -522,7 +522,7 @@ int ChimeraCcodeCommand::driver(linePair* filePos, string outputFName, string fi
 			delete candidateSeq;
 			
 			#if defined (__APPLE__) || (__MACH__) || (linux) || (__linux)
-				unsigned long int pos = inFASTA.tellg();
+				unsigned long long pos = inFASTA.tellg();
 				if ((pos == -1) || (pos >= filePos->end)) { break; }
 			#else
 				if (inFASTA.eof()) { break; }
@@ -547,7 +547,7 @@ int ChimeraCcodeCommand::driver(linePair* filePos, string outputFName, string fi
 }
 //**********************************************************************************************************************
 #ifdef USE_MPI
-int ChimeraCcodeCommand::driverMPI(int start, int num, MPI_File& inMPI, MPI_File& outMPI, MPI_File& outAccMPI, vector<unsigned long int>& MPIPos){
+int ChimeraCcodeCommand::driverMPI(int start, int num, MPI_File& inMPI, MPI_File& outMPI, MPI_File& outAccMPI, vector<unsigned long long>& MPIPos){
 	try {
 				
 		MPI_Status status;
