+++ screenseqscommand.h
@@ -33,9 +33,9 @@ public:
 private:
 
 	struct linePair {
-		unsigned long int start;
-		unsigned long int end;
-		linePair(unsigned long int i, unsigned long int j) : start(i), end(j) {}
+		unsigned long long start;
+		unsigned long long end;
+		linePair(unsigned long long i, unsigned long long j) : start(i), end(j) {}
 	};
 
 	vector<int> processIDS;   //processid
@@ -50,7 +50,7 @@ private:
 	int createProcesses(string, string, string, set<string>&);
 	
 	#ifdef USE_MPI
-	int driverMPI(int, int, MPI_File&, MPI_File&, MPI_File&, vector<unsigned long int>&, set<string>&);
+	int driverMPI(int, int, MPI_File&, MPI_File&, MPI_File&, vector<unsigned long long>&, set<string>&);
 	#endif
 
 	bool abort;
@@ -61,7 +61,7 @@ private:
 	map<string, int> nameMap;
 	int readNames();
 	
-	int getSummary(vector<unsigned long int>&);
+	int getSummary(vector<unsigned long long>&);
 	int createProcessesCreateSummary(vector<int>&, vector<int>&, vector<int>&, vector<int>&, vector<int>&, string);
 	int driverCreateSummary(vector<int>&, vector<int>&, vector<int>&, vector<int>&, vector<int>&, string, linePair*);	
 };
